requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.1
anyio==4.12.1
async-generator==1.10
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
beautifulsoup4==4.14.3
browser-use==0.11.8
browser-use-sdk==2.0.14
bubus==1.5.6
cdp-use==1.4.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
docstring_parser==0.17.0
fastuuid==0.14.0
filelock==3.21.2
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.62.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
inquirerpy==0.3.4
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.81.11
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
npyscreen==5.0.2
oauthlib==3.3.1
ollama==0.6.1
openai==2.16.0
outcome==1.3.0.post0
packaging==26.0
pfzy==0.3.4
pillow==12.1.0
playwright==1.58.0
portalocker==2.10.1
posthog==7.8.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
pyfiglet==1.0.4
Pygments==2.19.2
PyJWT==2.11.0
pyotp==2.9.0
pyparsing==3.3.2
pypdf==6.6.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
screeninfo==0.8.1
selenium==4.40.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sse-starlette==3.2.0
starlette==0.52.1
tavily-python==0.7.21
tenacity==9.1.3
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer==0.23.1
typer-slim==0.23.1
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid7==0.1.0
uvicorn==0.40.0
wcwidth==0.5.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
windows-curses==2.4.1
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
